Refactor GUI and Organize Utility Functions
- Removed unused imports in `utils.py`.
- Moved GUI utility functions from `gui_toolkit.py` to `utils.py`.
- Updated import statements in relevant files to use the new `utils.py`.
- Reorganized the code to improve code readability and maintainability.
- Ensured all screens are properly managed and displayed as expected.
