+++ src/subsearch/gui/screens/download_manager.py
@@ -50,6 +50,7 @@ class DownloadManager(tk.Frame):
         self.scrollbar.config(command=self.sub_listbox.yview)
         self.scrollbar.lift()
 
+            
     def fill_listbox(self) -> None:
         self.listbox_index: dict[int, Subtitle] = {}
         for enum, subtitle in enumerate(self.subtitles):
