+++ src/subsearch/gui/screens/language_options.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from tkinter import IntVar, ttk
 
 from subsearch.data.constants import FILE_PATHS
-from subsearch.gui import gui_toolkit
+from subsearch.gui import utils
 from subsearch.utils import io_toml
 
 
@@ -47,7 +47,7 @@ class SelectLanguage(ttk.Labelframe):
         providers = ", ".join(([_i.title() for _i in self.languages[toml_key]["incompatibility"]]))
         if providers:
             tip_text = f"If enabled, '{providers}' will be automatically skipped."
-            self.tip = gui_toolkit.ToolTip(btn, btn, tip_text)
+            self.tip = utils.ToolTip(btn, btn, tip_text)
             self.tip.show()
             self.tip_present = True
         btn.bind("<ButtonPress-1>", self.press_button)
